[Decomp][Object] Cohesion split: extract position/visibility to WorldObjectPosition.cpp
